+++ falcosidekick-k8s-operator/src/state.py
@@ -11,8 +11,13 @@ import ops
 from pydantic import BaseModel, ValidationError
 
 from config import CharmConfig, InvalidCharmConfigError
+from relations import InvalidLokiRelationError, LokiRelationManager
 
 logger = logging.getLogger(__name__)
+
+
+class InvalidStateError(Exception):
+    """Exception raised when a charm configuration is invalid and unrecoverable."""
 
 
 class CharmState(BaseModel):
@@ -23,12 +28,15 @@ class CharmState(BaseModel):
 
     Attributes:
         falcosidekick_listenport: The port on which Falcosidekick listens.
+        loki_endpoint_url: The URL of the Loki endpoint, if available.
     """
 
     falcosidekick_listenport: int
+    falcosidekick_loki_endpoint: str
+    falcosidekick_loki_hostport: str
 
     @classmethod
-    def from_charm(cls, charm: ops.CharmBase) -> "CharmState":
+    def from_charm(cls, charm: ops.CharmBase, loki_relation: LokiRelationManager) -> "CharmState":
         """Create a CharmState from a charm instance.
 
         Loads and validates the charm configuration, then constructs a CharmState
@@ -36,6 +44,7 @@ class CharmState(BaseModel):
 
         Args:
             charm: The charm instance from which to extract state.
+            loki_relation: The LokiRelationManager instance to get Loki relation data.
 
         Returns:
             CharmState: A validated CharmState instance.
@@ -45,13 +54,22 @@ class CharmState(BaseModel):
         """
         try:
             charm_config = charm.load_config(CharmConfig)
+            _url = loki_relation.get_loki_http_url()
+            loki_endpoint = _url.path if _url else "/loki/api/v1/push"
+            loki_hostport = f"{_url.scheme}://{_url.host}:{_url.port}" if _url else ""
         except ValidationError as e:
             logger.error("Configuration validation error: %s", e)
             error_fields = set(itertools.chain.from_iterable(err["loc"] for err in e.errors()))
             error_field_str = " ".join(f"{f}" for f in error_fields)
             raise InvalidCharmConfigError(f"Invalid charm configuration: {error_field_str}") from e
-
-        return cls(falcosidekick_listenport=charm_config.port)
+        except InvalidLokiRelationError as e:
+            logger.error("Loki relation data validation error: %s", e)
+            raise InvalidStateError("Invalid Loki relation data") from e
+        return cls(
+            falcosidekick_listenport=charm_config.port,
+            falcosidekick_loki_endpoint=loki_endpoint,
+            falcosidekick_loki_hostport=loki_hostport,
+        )
 
 
 class CharmBaseWithState(ops.CharmBase, ABC):