+++ docs/changelog.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 Each revision is versioned by the date of the revision.
 
+## 2025-12-29
+
+Add integration between falcosidekick-k8s-operator and loki-k8s-operator.
+
+### Added
+
+- Add relation between falcosidekick-k8s-operator and loki-k8s-operator for log forwarding.
+
 ## 2025-12-22
 
 Add port configuration option for falcosidekick-k8s-operator.